resolving PR comments removing 'required' and instead showing error message from back-end extracting form to separate component updating Cypress version to try get Semaphore testing to work
