+++ cypress/fixtures/body_blank_message.json
@@ -1,0 +1,3 @@
+{
+  "message": "Body can't be blank"
+}
