+++ cypress/fixtures/title_blank_message.json
@@ -1,0 +1,3 @@
+{
+  "message": "Title can't be blank"
+}
