+++ package.json
@@ -40,7 +40,7 @@
     ]
   },
   "devDependencies": {
-    "cypress": "^4.5.0",
+    "cypress": "^4.6.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "react-test-renderer": "^16.13.1"
